Refactor project in modules by functionality
Project refactoring split the project in 3 modules, one per
functionality: eventize, enrichment, dataframe processing.

Tests have been updated to use these new modules.

Tests have been also updated to allow their execution from
any path. CSV data files for tests are now loaded taking
into account their absolute path, obtained from test file
real path at runtime.
